minor fixes to prospects and products

+++ db/migrate/20170817151619_remove_business_units_supplier_table.rb
@@ -1,5 +1,7 @@
 class RemoveBusinessUnitsSupplierTable < ActiveRecord::Migration
   def change
-    drop_table :business_units_suppliers
+    if table exists? :business_units_suppliers
+      drop_table :business_units_suppliers
+    end
   end
 end
